parisc: Prevent overflow of PARISC_PDC_ENTRY_ORG

+++ src/parisc/machine-create.h
@@ -4,7 +4,7 @@ static struct machine_info CONCAT(machine_, MACHINE)  = {
 	.pdc_version = PARISC_PDC_VERSION,
 	.pdc_cpuid = PARISC_PDC_CPUID,
 	.pdc_caps = PARISC_PDC_CAPABILITIES,
-	.pdc_entry = PARISC_PDC_ENTRY_ORG,
+	.pdc_entry = (unsigned long) PARISC_PDC_ENTRY_ORG,
 	.pdc_cache_info = { PARISC_PDC_CACHE_INFO },
 	.device_list = { PARISC_DEVICE_LIST },
 };
